add compound stat type
This is only for internal use at this time. We're now gathering some
dense stats from /proc on Linux that would be really ugly to send upstream
one at a time, store, and query separately. They pretty much always want
to be used together, so optimize early.  This may change with experience.